ENSO Dynamical Systems Model
Project from:
 - Dynamical Systems
Languages:
 - Matlab
Description:
Sometimes systems don’t have behavior that we desire, and often there is chaos in system to consider. It is not a perfect world, but thankfully, we have a control feedback strategy that can steer the system toward particular equilibrium points. Sometimes we only need to apply one equation to control feedback, but it also can often be beneficial to apply it to both.  Chaotic behavior is difficult to study, and many times, it cannot be avoided. For example, the weather is very hard to predict. Model these models and control feedback can help the system become more stable.